+++ tests/unit/test_scheduler.py
@@ -5,6 +5,7 @@ from unittest.mock import patch
 
 import pytest
 
+from deode.host_actions import AmbigiousHostError, HostNotFoundError, SelectHost
 from deode.logs import logger
 from deode.scheduler import EcflowClient, EcflowServer, EcflowTask
 
@@ -43,27 +44,25 @@ class TestScheduler:
         ecf_port = ecflow_server._set_port_from_user(offset)
         assert port == ecf_port
 
-    def test_successfully_select_host_from_list(self, ecflow_server: EcflowServer):
+    def test_successfully_select_host_from_list(self):
         hostname = "localhost"
         hosts = ["foo", hostname]
-        ecf_host = ecflow_server._select_host_from_list(hosts)
+        ecf_host = SelectHost._select_host_from_list(hosts)
         assert ecf_host == hostname
 
-    def test_fail_to_find_host_from_list(self, ecflow_server: EcflowServer):
+    def test_fail_to_find_host_from_list(self):
         hosts = ["foo"]
         host_list = ",".join(hosts)
-        with pytest.raises(
-            RuntimeError, match=f"No ecflow host found, tried:{host_list}"
-        ):
-            ecflow_server._select_host_from_list(hosts)
+        with pytest.raises(HostNotFoundError, match=f"No host found, tried:{host_list}"):
+            SelectHost._select_host_from_list(hosts)
 
-    def test_to_many_hosts_found_from_list(self, ecflow_server: EcflowServer):
+    def test_to_many_hosts_found_from_list(self):
         hostname = "localhost"
         hosts = [hostname, hostname]
         host_list = ",".join(hosts)
         msg = f"Ambigious host selection:{host_list}"
-        with pytest.raises(RuntimeError, match=msg):
-            ecflow_server._select_host_from_list(hosts)
+        with pytest.raises(AmbigiousHostError, match=msg):
+            SelectHost._select_host_from_list(hosts)
 
     def test_ecflow_client(self, ecflow_server: EcflowServer, ecflow_task):
         EcflowClient(ecflow_server, ecflow_task)